added check for completed chromosomes

+++ impute_INGI_ref_array.sh
@@ -215,8 +215,13 @@ for chunk in `seq 1 $chunk_num`; do
 	fi
 done
 
-#delete duplicate lines from the command file
-awk '!_[$0]++' $imputedir/chr${chr}_command.list.tmp > $imputedir/chr${chr}_command.list
-rm $imputedir/chr${chr}_command.list.tmp
+#check if tmp command file exists before lauching new jobs
+if [[ -s $imputedir/chr${chr}_command.list.tmp ]]; then
+	#delete duplicate lines from the command file
+	awk '!_[$0]++' $imputedir/chr${chr}_command.list.tmp > $imputedir/chr${chr}_command.list
+	rm $imputedir/chr${chr}_command.list.tmp
 
-mkdir -p $imputedir/LOGS;size=`wc -l $imputedir/chr${chr}_command.list|cut -f 1 -d " "`;bsub -J "${PANEL}${postfix}.${pop}.chr${chr}[1-${size}]" -q $queue -R "select[mem>$mem] rusage[mem=$mem]" -M${mem} -o "$imputedir/LOGS/%J_${PANEL}${postfix}.${pop}.chr${chr}.%I.log" -e "$imputedir/LOGS/%J_${PANEL}${postfix}.${pop}.chr${chr}.%I.err" -- ~/Work/bash_scripts/ja_runner_par.sh -l $imputedir/chr${chr}_command.list
+	mkdir -p $imputedir/LOGS;size=`wc -l $imputedir/chr${chr}_command.list|cut -f 1 -d " "`;bsub -J "${PANEL}${postfix}.${pop}.chr${chr}[1-${size}]" -q $queue -R "select[mem>$mem] rusage[mem=$mem]" -M${mem} -o "$imputedir/LOGS/%J_${PANEL}${postfix}.${pop}.chr${chr}.%I.log" -e "$imputedir/LOGS/%J_${PANEL}${postfix}.${pop}.chr${chr}.%I.err" -- ~/Work/bash_scripts/ja_runner_par.sh -l $imputedir/chr${chr}_command.list
+else
+	echo "Chromosome ${chr} already COMPLETED!!job not submitted!! "
+fi
